refactors and cleaned models
