+++ XYZAssociation/nbproject/private/private.properties
@@ -1,10 +1,10 @@
-deploy.ant.properties.file=C:\\Users\\j3-lui.CAMPUS\\AppData\\Roaming\\Netbeans\\8.2\\config\\GlassFishEE6\\Properties\\gfv3332573768.properties
+deploy.ant.properties.file=C:\\Users\\jason\\AppData\\Roaming\\NetBeans\\8.2\\config\\GlassFishEE6\\Properties\\gfv3803245337.properties
 j2ee.platform.is.jsr109=true
-j2ee.server.domain=C:/Users/j3-lui.CAMPUS/AppData/Roaming/Netbeans/8.2/config/GF_4.1.1/domain1
-j2ee.server.home=C:/Program Files (x86)/glassfish-4.1.1/glassfish
-j2ee.server.instance=[C:\\Program Files (x86)\\glassfish-4.1.1\\glassfish;C:\\Program Files (x86)\\glassfish-4.1.1\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
-j2ee.server.middleware=C:/Program Files (x86)/glassfish-4.1.1
+j2ee.server.domain=F:/glassfish-4.1.1/glassfish/domains/domain1
+j2ee.server.home=F:/glassfish-4.1.1/glassfish
+j2ee.server.instance=[F:\\glassfish-4.1.1\\glassfish;F:\\glassfish-4.1.1\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.middleware=F:/glassfish-4.1.1
 javac.debug=true
 javadoc.preview=true
 selected.browser=Chrome
-user.properties.file=C:\\Users\\j3-lui.CAMPUS\\AppData\\Roaming\\Netbeans\\8.2\\build.properties
+user.properties.file=C:\\Users\\jason\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
