+++ XYZAssociation/nbproject/private/private.xml
@@ -3,22 +3,31 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/WEB-INF/web.xml</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/registrationView.jsp</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/MakePaymentController.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/makeClaim.jsp</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/User.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/AuthenticatorBean.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/ShowMemberInfo.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/RegisterBean.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/makePayment.jsp</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/GetMemberClaimsBean.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/loginView.jsp</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/CheckCookie.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/memberInfo.jsp</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/RegistrationController.java</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/web/index.html</file>
-            <file>file:/D:/Github%20Archive/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/AuthenticateUser.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/DBConnectorBean.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/Member.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/DBClaimINSERT.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/registrationView.jsp</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/Payment.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/submitClaim.jsp</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/DBAuthenticatorBean.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/SubmitClaimController.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/DBRegisterBean.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/noLoginView.jsp</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/RegisterBean.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/User.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/Claim.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/ShowMemberInfo.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/AuthenticateUser.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/DBMemberRETURN.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/WEB-INF/web.xml</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/model/DBClaimRETURN.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/claimsAndPayments.jsp</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/makePayment.jsp</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/MakeClaimController.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/RegistrationController.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/src/java/controller/ShowClaimsAndPayments.java</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/nbproject/build-impl.xml</file>
+            <file>file:/F:/bigTimeESD/XYZAssociation/WebApplication1/web/loginView.jsp</file>
         </group>
     </open-files>
 </project-private>
